refactor(agent-hooks): split hooks-JSON file I/O and SFTP primitives into modules
Mechanical extraction, no behavior change: the hooks settings JSON contract
(types + read/write/retry primitives) moves to hooks-json-file.ts and the
timeout-guarded ssh2 wrappers move to sftp-file-primitives.ts. Restores the
max-lines budget the concurrency guard exceeded; installer-utils re-exports
the JSON surface so call sites keep their single historical import.
