+++ src/main/agent-hooks/hooks-json-read.ts
@@ -1,5 +1,5 @@
 import { existsSync, readFileSync } from 'node:fs'
-import type { HooksConfig } from './installer-utils'
+import type { HooksConfig } from './hooks-json-file'
 
 export function isPlainObject(value: unknown): value is Record<string, unknown> {
   return typeof value === 'object' && value !== null && !Array.isArray(value)
